requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
SQLAlchemy==2.0.39
psycopg2-binary==2.9.10
pydantic==2.10.6
alembic==1.15.1
python-dotenv==1.0.1
celery==5.4.0
redis==5.2.1
